Time: 9 ms (47.26%), Space: 8.3 MB (48.18%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -6,7 +6,7 @@ public:
         std::cout.tie(nullptr);
 
         if(x < 0) return false;
-        long res = 0, y = x;
+        unsigned int res = 0, y = x;
         
         while(y) {
             res *= 10;
